+++ src/updateKeploy.ts
@@ -1,110 +1,105 @@
 import * as vscode from 'vscode';
+import {exec} from 'child_process';
+import * as os from 'os';
+import getKeployVersion from './version';
+const execShell = (cmd: string) => {
+    return new Promise<string>((resolve, reject) => {
+        let commandToExecute: string;
+        if (os.platform() === 'win32') {
+            // Execute the command in the WSL environment if on Windows
+            commandToExecute = `wsl ${cmd}`;
+        } else {
+            // Execute the command in the default shell if on other platforms
+            commandToExecute = cmd;
+        }
 
-export async function downloadAndUpdate(downloadUrl: string) {
-    let aliasPath = "/usr/local/bin/keploybin"; 
-    try {
-      let bashPath: string;
-      if (process.platform === 'win32') {
-          // If on Windows, use the correct path to WSL's Bash shell
-          bashPath = 'wsl.exe';
-      } else {
-          // Otherwise, assume Bash is available at the standard location
-          bashPath = '/bin/bash';
-      }
-        // Create a new terminal instance with the Bash shell
-        const terminal = vscode.window.createTerminal({
-            name: 'Keploy Terminal',
-            shellPath: bashPath
+        exec(commandToExecute, (err, stdout, stderr) => {
+            if (err) {
+                reject(err.message);
+                return;
+            }
+            if (stderr) {
+                reject(stderr);
+                return;
+            }
+            resolve(stdout);
         });
-
-        // Show the terminal
-        terminal.show();
-
-        const curlCmd = `curl --silent --location ${downloadUrl} | tar xz -C /tmp `;
-        const moveCmd = `sudo mv /tmp/keploy ${aliasPath}`;
-        terminal.sendText(curlCmd);
-        terminal.sendText(moveCmd);
-        // Display an information message
-        vscode.window.showInformationMessage('Downloading and updating Keploy binary...');
-    } catch (error:any) {
-        console.error('Error updating Keploy binary:', error.message);
+    });
+};
+export async function downloadAndUpdate(downloadUrl: string): Promise<void> {
+    const output = await execShell('keploy --version');
+    const keployIndex = output.indexOf('Keploy');
+    let keployVersion = '';
+    if (keployIndex !== -1) {
+        keployVersion = output.substring(keployIndex + 'Keploy'.length).trim();
     }
+
+    console.log('Current Keploy version:', keployVersion);
+    if (!keployVersion) {
+        vscode.window.showErrorMessage('You dont have Keploy installed. Please install Keploy first');
+        return;
+    }
+    const latestVersion = await getKeployVersion();
+    // Remove "v" from the beginning of the latest version string, if present
+    const formattedLatestVersion = latestVersion.startsWith('v') ? latestVersion.substring(1) : latestVersion;
+    
+    console.log('Latest Keploy version:', formattedLatestVersion);
+
+    if (keployVersion === formattedLatestVersion) {
+        vscode.window.showInformationMessage('Keploy is already up to date');
+        return;
+    }
+
+     console.log('Downloading and updating Keploy binary...');
+     vscode.window.showInformationMessage('Downloading and updating Keploy binary...');
+    return new Promise<void>((resolve, reject) => {
+        
+        let aliasPath = "/usr/local/bin/keploybin"; 
+        try {
+            let bashPath: string;
+            if (process.platform === 'win32') {
+                // If on Windows, use the correct path to WSL's Bash shell
+                bashPath = 'wsl.exe';
+            } else {
+                // Otherwise, assume Bash is available at the standard location
+                bashPath = '/bin/bash';
+            }
+            // Create a new terminal instance with the Bash shell
+            const terminal = vscode.window.createTerminal({
+                name: 'Keploy Terminal',
+                shellPath: bashPath
+            });
+
+            // Show the terminal
+            terminal.show();
+
+            const curlCmd = `curl --silent --location ${downloadUrl} | tar xz -C /tmp `;
+            const moveCmd = `sudo mv /tmp/keploy ${aliasPath}`;
+
+            // Execute commands asynchronously
+            const executeCommand = async (command: string) => {
+                return new Promise<void>((resolve, reject) => {
+                    terminal.sendText(command);
+                    // terminal.sendText('echo $?'); // Output the exit status
+                    // (window as any).onDidWriteTerminalData((event: any) => console.log(event.data.trim()))
+                    setTimeout(() => resolve(), 8000); // Assuming commands will complete within 3 seconds
+                    //we need to figure out a way to find out if command is completed successfully or not
+                });
+            };
+
+            // Execute commands sequentially
+            Promise.all([
+                executeCommand(curlCmd),
+                executeCommand(moveCmd)
+            ]).then(() => {
+                // Display an information message
+                vscode.window.showInformationMessage('Downloading and updating Keploy binary...');
+                resolve(); // Resolve the promise if all commands succeed
+            }).catch(error => {
+                reject(error); // Reject the promise if any command fails
+            });
+        } catch (error) {
+            reject(error); // Reject the promise if an error occurs during execution
+        }
+    });
 }
-
-
-// import util from 'util';
-// import { exec, spawn, spawnSync } from 'child_process';
-// import os from 'os';
-// import fs from 'fs';
-// const execPromise = util.promisify(exec);
-
-// export async function downloadAndUpdate(downloadUrl: string) {
-
-
-  
-// //       // Download keploy binary
-// //       
-// //       // const curlCmd = ` curl -O https://raw.githubusercontent.com/keploy/keploy/main/keploy.sh && source keploy.sh`;
-// //       // const curlCmd = `curl -O https://raw.githubusercontent.com/keploy/keploy/main/keploy.sh && bash keploy.sh`;
-// //       // const untarCmd = 'tar xz -C /tmp';
-// //       // let child = spawnSync("wsl", [ "bash", "-l", "-c", curlCmd ], {
-// //         // shell: true
-// //       // });
-// //       // const downloadProcess = exec(curlCmd, (error, stdout, stderr) => {
-// //       //   if (error) {
-// //       //     console.error(`exec error: ${error}`);
-// //       //     return;
-// //       //   }
-// //       //   console.log(`stdout: ${stdout}`);
-// //       //   console.error(`stderr: ${stderr}`);
-// //       // });
-// //     //   const untarProcess = exec(untarCmd);
-// //     // console.log(child.stdout.toString())
-// //     // console.log(child.stderr.toString())
-      
-// //     //   // Pipe the output of the first command to the second command
-// //     //   if (!downloadProcess.stdout || !untarProcess.stdin) {
-// //     //     throw new Error('Failed to pipe commands');
-// //     //   }
-// //     //   downloadProcess.stdout.pipe(untarProcess.stdin);
-  
-// //     //   await Promise.all([
-// //     //     new Promise<void>((resolve, reject) => {
-// //     //         downloadProcess.on('exit', (code:any) => {
-// //     //             if (code === 0) {
-// //     //                 resolve(undefined);
-// //     //             } else {
-// //     //                 reject(new Error(`Download command failed with code ${code}`));
-// //     //             }
-// //     //         });
-// //     //     }),
-// //     //     new Promise((resolve, reject) => {
-// //     //       untarProcess.on('exit', (code:any) => {
-// //     //         if (code === 0) {
-// //     //           resolve(undefined);
-// //     //         } else {
-// //     //           reject(new Error(`Untar command failed with code ${code}`));
-// //     //         }
-// //     //       });
-// //     //     }),
-// //     //   ]);
-  
-// //     //   // Move keploy binary to aliasPath
-
-// //     //   await execPromise(moveCmd);
-// //     //   await new Promise((resolve, reject) => {
-// //     //     child.on('exit', (code:any) => {
-// //     //       if (code === 0) {
-// //     //         resolve(undefined);
-// //     //       } else {
-// //     //         reject(new Error(`Download command failed with code ${code}`));
-// //     //       }
-// //     //     });
-// //     //   });
-  
-// //       console.log('Keploy binary updated successfully.');
-// //     } catch (error:any) {
-// //       console.error('Error updating Keploy binary:', error.message);
-// //     }
-// //   }
-